requirements.txt
# Core Dependencies
feedparser>=6.0.10
python-telegram-bot>=20.0
requests>=2.31.0
sqlalchemy>=2.0.0
apscheduler>=3.10.0
python-dateutil>=2.8.2
python-dotenv>=1.0.0
pyyaml>=6.0
openai>=1.0.0
telegraph>=2.2.0

# ML (lightweight, scikit-learn is sufficient for our use case)
scikit-learn>=1.3.0

# Development (optional, only install for development)
# pytest>=7.4.0
# black>=23.0.0
# pylint>=2.17.0
